// backend/src/model/dto/books.dto.ts
import { Book } from "../entities/books/book";
import { PointType } from "../entities/points/point_type";
import { ReadingPoint } from "../entities/points/reading_point";

export class PopularBooksFilterDto {
    facultyId?: number;
    pointId?: number;
    universityWide?: boolean;
}

export class BookPopularityDto {
    bookId: number;
    title: string;
    author: string;
    totalOrders: number;
}

export class BookStatsFilterDto {
    pointId?: number;
    userId?: number;
    author?: string;
    releaseYear?: number;
    admissionYear?: number;
    libraryWide?: boolean;
}

export class BookStatResultDto {
    bookId: number;
    title: string;
    author: string;
    releaseDate: Date;
    admissionDate: Date;
    lostDate?: Date;
    status: 'Поступила' | 'Утеряна';
}

export class BookStatsResponseDto {
    books: BookStatResultDto[];
    totalReceived: number;
    totalLost: number;
}

export class BookWithReadingPointDto {
    bookId: number;
    title: string;
    author: string;
    releaseDate: Date;
    pointId: number;
    type: PointType;
    address: string;
}